+++ hardware/Dummy.cpp
@@ -131,7 +131,7 @@ namespace http {
 			int sensortype = atoi(ssensortype.c_str());
 			unsigned int type = 0;
 			unsigned int subType = 0;
-			uint64_t DeviceRowIdx = -1;
+			uint64_t DeviceRowIdx = (uint64_t )-1;
 
 			for (int i = 0; i < sizeof(mappedsensorname) / sizeof(mappedsensorname[0]); i++)
 			{
